+++ home/tlh/default.nix
@@ -55,9 +55,6 @@
       outputs.overlays.additions
       outputs.overlays.modifications
       outputs.overlays.stable-packages
-      outputs.overlays.f2k-packages
-      outputs.overlays.chaotic-packages
-      inputs.nixpkgs-f2k.overlays.default
       inputs.nur.overlays.default
     ];