+++ home/user/default.nix
@@ -48,8 +48,6 @@
       outputs.overlays.additions
       outputs.overlays.modifications
       outputs.overlays.stable-packages
-      outputs.overlays.f2k-packages
-      inputs.nixpkgs-f2k.overlays.default
       inputs.nur.overlays.default
     ];