+++ hosts/matangi/default.nix
@@ -13,10 +13,7 @@
     outputs.overlays.additions
     outputs.overlays.modifications
     outputs.overlays.stable-packages
-    outputs.overlays.master-packages
-    outputs.overlays.f2k-packages
-    outputs.overlays.chaotic-packages
-    inputs.antigravity-nix.overlays.default
+    inputs.nur.overlays.default
   ];
 
   modules = {