+++ modules/nixos/base/nix.nix
@@ -129,16 +129,11 @@ in {
           [
             "https://cache.garnix.io"
             "https://cache.nixos.org?priority=10"
-            "https://fortuneteller2k.cachix.org"
             "https://nix-community.cachix.org"
             "https://nixpkgs-unfree.cachix.org"
             "https://pre-commit-hooks.cachix.org"
             "https://cuda-maintainers.cachix.org"
             "https://ai.cachix.org"
-            "https://stable-diff.cachix.org"
-            "https://sanatanalinux.cachix.org"
-            "https://chaotic-nyx.cachix.org"
-            "https://zen-browser.cachix.org"
             "https://nix-gaming.cachix.org"
           ]
           ++ cfg.extraCaches;
@@ -148,16 +143,11 @@ in {
           [
             "cache.garnix.io:CTFPyKSLcx5RMJKfLo5EEPUObbA78b0YQ2DTCJXqr9g="
             "cache.nixos.org-1:6NCHdD59X431o0gWypbMrAURkbJ16ZPMQFGspcDShjY="
-            "fortuneteller2k.cachix.org-1:kXXNkMV5yheEQwT0I4XYh1MaCSz+qg72k8XAi2PthJI="
             "nix-community.cachix.org-1:mB9FSh9qf2dCimDSUo8Zy7bkq5CX+/rkCWyvRCYg3Fs="
             "nixpkgs-unfree.cachix.org-1:hqvoInulhbV4nJ9yJOEr+4wxhDV4xq2d1DK7S6Nj6rs="
             "pre-commit-hooks.cachix.org-1:Pkk3Panw5AW24TOv6kz3PvLhlH8puAsJTBbOPmBo7Rc="
             "cuda-maintainers.cachix.org-1:0dq3bujKpuEPMCX6U4WylrUDZ9JyUG0VpVZa7CNfq5E="
             "ai.cachix.org-1:N9dzRK+alWwoKXQlnn0H6aUx0lU/mspIoz8hMvGvbbc="
-            "stable-diff.cachix.org-1:liYFm3f3q1dAoilj2Ag2IEKzW3Q9/HJcLlrAIytAcy0="
-            "sanatanalinux.cachix.org-1:9WsJYECJ+Lt0HPTUI7+6f9uAaAUouaBUyTd9iAJbUEY="
-            "chaotic-nyx.cachix.org-1:HfnXSw4pj95iI/n17rIDy40agHj12WfF+Gqk6SonIT8="
-            "zen-browser.cachix.org-1:VkBZFD6ielaF2pp1M0KGDAXdYqB25lX5x0bJ8uMnbHs="
             "nix-gaming.cachix.org-1:nbjlureqMbRAxR1gJ/f3hxemL9svXaZF/Ees8vCUUs4="
           ]
           ++ cfg.extraTrustedKeys;